Added the ability to schedule an index to happen in the future

+++ src/data/search/data-source.js
@@ -1,6 +1,8 @@
 import { dataSource as CoreDataSource } from '@apollosproject/data-connector-algolia-search';
 import { createGlobalId } from '@apollosproject/server-core';
 import ApollosConfig from '@apollosproject/config'
+import Queue from 'bull'
+import Redis from 'ioredis';
 import { graphql } from 'graphql';
 import { take, get } from 'lodash';
 import sanitizeHtml from 'sanitize-html';
@@ -11,6 +13,32 @@ import momentTz from 'moment-timezone'
 const { ROCK } = ApollosConfig
 
 const MAX_SIZE = 10000 // bytes
+const { REDIS_URL, CONTENT } = process.env;
+
+const redis = new Redis(REDIS_URL);
+let client;
+let subscriber;
+let queueOpts;
+
+if (REDIS_URL) {
+  client = new Redis(REDIS_URL);
+  subscriber = new Redis(REDIS_URL);
+
+  // Used to ensure that N+3 redis connections are not created per queue.
+  // https://github.com/OptimalBits/bull/blob/develop/PATTERNS.md#reusing-redis-connections
+  queueOpts = {
+    createClient(type) {
+      switch (type) {
+        case 'client':
+          return client;
+        case 'subscriber':
+          return subscriber;
+        default:
+          return new Redis(REDIS_URL);
+      }
+    },
+  };
+}
 
 const cleanHtmlContentForIndex = (htmlContent) => {
   // Strip all html tags
@@ -59,6 +87,29 @@ const processObjectSize = (obj) => {
   }
 }
 
+const deleteKeysByPattern = (pattern) => {
+  return new Promise((resolve, reject) => {
+    const stream = redis.scanStream({
+      match: pattern
+    });
+    stream.on("data", (keys) => {
+      if (keys.length) {
+        const pipeline = redis.pipeline();
+        keys.forEach((key) => {
+          pipeline.del(key);
+        });
+        pipeline.exec();
+      }
+    });
+    stream.on("end", () => {
+      resolve();
+    });
+    stream.on("error", (e) => {
+      reject(e);
+    });
+  });
+};
+
 export default class Search extends CoreDataSource {
 
   async indexAllGeneralContent() {
@@ -78,30 +129,90 @@ export default class Search extends CoreDataSource {
   }
 
   async updateContentItemIndex(id) {
-    // Resolve the Content Item
+    const log = (msg) => console.log(`\n\x1b[35m${msg}\x1b[30m\n`)
+    /** Resolve the Content Item */
     const { ContentItem } = this.context.dataSources
     const item = await ContentItem.getFromId(id)
+    if (!item) return null
+
     const hideFromSearch = get(item, "attributeValues.hideFromSearch.value", "false").toLowerCase()
 
-    if (hideFromSearch === "true") { // Delete the item if it should not be included in Search
+    /** Delete the item if it should not be included in Search */
+    if (hideFromSearch === "true") {
       const type = await this.resolveContentItem(item);
+      return
       return this.index.deleteObject(createGlobalId(item.id, type))
     }
 
+    /** Resolve the item to an indexable item and get the Start Date */
+    const indexableItem = await this.mapItemToAlgolia(item)
     const { startDateTime } = item
 
+    /** Set a key for the index that is unique to the item.
+     *  
+     *  Since Rock is telling us that this Content Item needs
+     *  to be update, we need to delete all currently existing
+     *  jobs to make sure that we don't index more than needed
+     */
+    const indexKey = `algolia-index-${indexableItem.id}`
+    await deleteKeysByPattern(`bull:${indexKey}:*`)
+
+    /** If a start date exists, and that date is in the future, we want to
+     *  set a job that updates the item automatically on the day it should
+     *  be active
+     */
     if (startDateTime && startDateTime !== '') {
       const mStartDateTime = momentTz(startDateTime).tz(ROCK.TIMEZONE)
       if (mStartDateTime.isValid() && mStartDateTime.isAfter(momentTz())) {
-        console.log("SET A JOB")
+        /** Set up the options for the bull job.
+         * 
+         *  We need to get the difference between now and the Start Time
+         *  in milliseconds to set the delay. We can keep the attempts at
+         *  2 just to be safe, but we shouldn't see any issues.
+         * 
+         *  Our Prefix gets set to include our content type just so that
+         *  Redis doesn't get confused between our state and production
+         *  content types.
+         * 
+         *  Since we're already in communication with Rock, let's just
+         *  resolve the item to the indexable item and save that as a
+         *  part of our data
+         */
+        const itemQueue = new Queue(indexKey, REDIS_URL)
+        const data = {
+          action: 'update',
+          item: indexableItem,
+          timestamp: momentTz().format('hh:mm:ss')
+        };
+        const options = {
+          delay: mStartDateTime.diff(momentTz()),
+          attempts: 2,
+          prefix: `bull-${CONTENT}`,
+        };
 
-        console.log(this)
+        log(`Scheduling search index update for "${item.title}"`)
+        itemQueue.add(data, options);
+        itemQueue.process(job => {
+          /** Get the item from our job data so that we can go ahead
+           *  and execute our search
+           * 
+           *  TODO : end date auto-remove
+           */
+          const { data } = job
+          const { action, item } = data
+
+          if (action === "update") {
+            log(`Running scheduled search index update for "${item.title}"`)
+            return this.addObjects([item])
+          }
+        });
+
+        return null
       }
     }
 
-    const indexableItem = await this.mapItemToAlgolia(item)
-
-    // return this.addObjects([indexableItem])
+    log(`Updating search index for "${item.title}"`)
+    return this.addObjects([indexableItem])
   }
 
   resolveContentItem(item) {
